phone-fix-backend/requirements.txt
Django>=5.0,<6.0
strawberry-graphql-django
psycopg2-binary
django-environ
django-cors-headers
PyJWT>=2.8.0
